Fix dtypes

+++ src/correct_images/tools/numerical.py
@@ -184,7 +184,7 @@ def mean_std(data, calculate_std=True):
 
 
 @nb.jit(
-    nb.types.UniTuple(nb.float64, 2)(nb.float64[:], nb.int64, nb.float64, nb.int64),
+    nb.types.UniTuple(nb.float32, 2)(nb.float32[:], nb.int64, nb.float64, nb.int64),
     fastmath=True,
     nopython=True,
 )
@@ -217,7 +217,7 @@ def ransac_mean_std_1d(data, max_iterations=1000, threshold=0.1, sample_size=2):
         mean = np.mean(data[idxs])
         # Compute the inliers given the threshold
         inliers_idx = np.argwhere(np.abs(data - mean) < threshold).flatten()
-        inliers = np.zeros((len(inliers_idx)))
+        inliers = np.zeros((len(inliers_idx)), dtype=data.dtype)
         for i, idx in enumerate(inliers_idx):
             inliers[i] = data[idx]
         # Compute the error
